backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
scikit-learn
pydantic
python-dotenv
PyPDF2